requirements.txt
django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1
celery==5.5.3
redis==6.2.0
python-decouple==3.8
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
gunicorn==23.0.0